% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan2.R
\name{newPlan}
\alias{newPlan}
\title{Get linearly scaled variable based on optimal bins}
\description{
Finds outliers and then bands between 0 and 1 on optimal bins of
non-outlier data
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-newPlan-new}{\code{newPlan$new()}}
\item \href{#method-newPlan-viewGantt}{\code{newPlan$viewGantt()}}
\item \href{#method-newPlan-viewEV}{\code{newPlan$viewEV()}}
\item \href{#method-newPlan-clone}{\code{newPlan$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newPlan-new"></a>}}
\if{latex}{\out{\hypertarget{method-newPlan-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newPlan$new(
  phase,
  activity,
  start,
  due,
  progress,
  planned_cost,
  project_value,
  cost_to_date,
  date = today()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newPlan-viewGantt"></a>}}
\if{latex}{\out{\hypertarget{method-newPlan-viewGantt}{}}}
\subsection{Method \code{viewGantt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newPlan$viewGantt()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newPlan-viewEV"></a>}}
\if{latex}{\out{\hypertarget{method-newPlan-viewEV}{}}}
\subsection{Method \code{viewEV()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newPlan$viewEV()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-newPlan-clone"></a>}}
\if{latex}{\out{\hypertarget{method-newPlan-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{newPlan$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
