% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan2.R
\name{plot,newPlan-method}
\alias{plot,newPlan-method}
\title{Get linearly scaled variable based on optimal bins}
\usage{
\S4method{plot}{newPlan}(x)
}
\description{
Finds outliers and then bands between 0 and 1 on optimal bins of
non-outlier data
}
